requirements.txt
pandas>=2.1.0
numpy>=2.0.0
scikit-learn>=1.5.0
joblib>=1.4.0
tabulate>=0.9.0
pytz
playwright>=1.40.0
pytest-playwright